dissolve "custom extensions" test into proper files
